\name{CalcAgeMaxFert}
\alias{CalcAgeMaxFert}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcAgeMaxFert (beta, modelFert = "M1", ageMatur = 0, maxAge = 100)}

\arguments{
	\item{ beta }{FILL UP}

	\item{modelFert }{FILL UP}

	\item{ageMatur }{FILL UP}

	\item{maxAge }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

