\name{CalcAveDemo}
\alias{CalcAveDemo}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcAveDemo (demo)}

\arguments{
	\item{ demo }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{}

\keyword{FILL UP}

