\name{CalcDemo}
\alias{CalcDemo}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcDemo (theta = NULL, beta = NULL, x = NULL, dx = NULL, model = "GO",  shape = "simple", modelFert = "M1", type = "both", minSx = 0.01,  summarStats = TRUE, ageMatur = 0, maxAge = NULL, agesAR = NULL,  SxValsAR = NULL)}

\arguments{
	\item{ theta }{FILL UP}

	\item{beta }{FILL UP}

	\item{x }{FILL UP}

	\item{dx }{FILL UP}

	\item{model }{FILL UP}

	\item{ shape }{FILL UP}

	\item{modelFert }{FILL UP}

	\item{type }{FILL UP}

	\item{minSx }{FILL UP}

	\item{ summarStats }{FILL UP}

	\item{ageMatur }{FILL UP}

	\item{maxAge }{FILL UP}

	\item{agesAR }{FILL UP}

	\item{ SxValsAR }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

