\name{Calcfert}
\alias{Calcfert}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{Calcfert (beta, x, modelFert = "M1", checkBeta = TRUE)}

\arguments{
	\item{ beta }{FILL UP}

	\item{x }{FILL UP}

	\item{modelFert }{FILL UP}

	\item{checkBeta }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

