\name{CalcKaplanMeier}
\alias{CalcKaplanMeier}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcKaplanMeier (ageLast, departType)}

\arguments{
	\item{ ageLast }{FILL UP}

	\item{departType }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

