\name{CalcKaplanMeierCIs}
\alias{CalcKaplanMeierCIs}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcKaplanMeierCIs (ageLast, departType, alpha = 0.05, nboot = 1000)}

\arguments{
	\item{ ageLast }{FILL UP}

	\item{departType }{FILL UP}

	\item{alpha }{FILL UP}

	\item{nboot }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

