\name{CalcLifeTableCIs}
\alias{CalcLifeTableCIs}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcLifeTableCIs (ageLast, ageFirst = NULL, departType, nboot = 2000,  alpha = 0.05)}

\arguments{
	\item{ ageLast }{FILL UP}

	\item{ageFirst }{FILL UP}

	\item{departType }{FILL UP}

	\item{nboot }{FILL UP}

	\item{ alpha }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

