\name{CalcProductLimitEstCIs}
\alias{CalcProductLimitEstCIs}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcProductLimitEstCIs (ageFirst, ageLast, departType, nboot = 1000, alpha = 0.05)}

\arguments{
	\item{ ageFirst }{FILL UP}

	\item{ageLast }{FILL UP}

	\item{departType }{FILL UP}

	\item{nboot }{FILL UP}

	\item{alpha }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

