\name{CalcSurvProbs}
\alias{CalcSurvProbs}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcSurvProbs (demo, dx = 1)}

\arguments{
	\item{ demo }{FILL UP}

	\item{dx }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

