\name{FindSilerPars}
\alias{FindSilerPars}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{FindSilerPars (theta, palpha)}

\arguments{
	\item{ theta }{FILL UP}

	\item{palpha }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

