\name{myoptim}
\alias{myoptim}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{myoptim (par, fn, ...)}

\arguments{
	\item{ par }{FILL UP}

	\item{fn }{FILL UP}

	\item{... }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

