\name{paramDemo-package}
\alias{paramDemo-package}
\alias{paramDemo}

\docType{package}

\title{
 FILL UP
}

\description{
 FILL UP
}

\details{
\tabular{ll}{
	Package: \tab paramDemo \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2021-10-25\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}

FILL UP DETAILS 
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\references{
FILL UP
}