\name{plot.lifeTableCIs}
\alias{plot.lifeTableCIs}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{plot.lifeTableCIs (x, demorate = "lx", ...)}

\arguments{
	\item{ x }{FILL UP}

	\item{demorate }{FILL UP}

	\item{... }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

