\name{plot.paramDemo}
\alias{plot.paramDemo}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{plot.paramDemo (x, demofun = "all", ...)}

\arguments{
	\item{ x }{FILL UP}

	\item{demofun }{FILL UP}

	\item{... }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

