\name{SampleRandAge}
\alias{SampleRandAge}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{SampleRandAge (n, theta, dx = 0.001, model = "GO", shape = "simple",  minSx = 1e-04)}

\arguments{
	\item{ n }{FILL UP}

	\item{theta }{FILL UP}

	\item{dx }{FILL UP}

	\item{model }{FILL UP}

	\item{shape }{FILL UP}

	\item{ minSx }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

