\name{SimulateStudy}
\alias{SimulateStudy}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{SimulateStudy (n, theta, model, shape, studyStart, studySpan, birthStart,  birthUncert = 0, dx = 1/365.25, provideSurv = TRUE, checkTheta = TRUE)}

\arguments{
	\item{ n }{FILL UP}

	\item{theta }{FILL UP}

	\item{model }{FILL UP}

	\item{shape }{FILL UP}

	\item{studyStart }{FILL UP}

	\item{studySpan }{FILL UP}

	\item{birthStart }{FILL UP}

	\item{ birthUncert }{FILL UP}

	\item{dx }{FILL UP}

	\item{provideSurv }{FILL UP}

	\item{checkTheta }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

