\name{CalcLifeTable}
\alias{CalcLifeTable}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcLifeTable (ageLast, ageFirst = NULL, departType)}

\arguments{
	\item{ ageLast }{FILL UP}

	\item{ageFirst }{FILL UP}

	\item{departType }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

