\name{CalcSurv}
\alias{CalcSurv}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{CalcSurv (theta, x, model = "GO", shape = "simple", checkTheta = TRUE)}

\arguments{
	\item{ theta }{FILL UP}

	\item{x }{FILL UP}

	\item{model }{FILL UP}

	\item{shape }{FILL UP}

	\item{checkTheta }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

