\name{paramDemo-package}
\alias{paramDemo-package}
\alias{paramDemo}

\docType{package}

\title{
 Parametric and Non-Parametric Demographic Functions and Applications
}

\description{
 Functions for parametric and non-parametric calculations of survival, mortality and fertility
}

\details{
\tabular{ll}{
	Package: \tab paramDemo \cr
	Type: \tab Package\cr
	Version: \tab 1.0.1\cr
	Date: \tab 2024-07-03\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}
This package contains functions for calculations of parametric mortality, survival and fertility models, as well as non-parametric methods for life tables and product limit estimators.
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}
